Feat qol

+++ src/lambdas/lambda_delete_file.py
@@ -117,30 +117,34 @@ def remove_single_file(username: str, album_uuid: str, file_uuid: str):
 
 def remove_album(username: str, owner_album_uuid: str, album_uuid: str):
     remove_sharing_instances(username, album_uuid)
+
+    all_owners = get_all_owners(username, album_uuid)
+
+    if len(all_owners) == 1: # If this file is only in 1 album.
+        
+        # Remove each subfile: THIS HAS TO GO BEFORE BECAUSE:
+        # if you remove [user-album], then when you delete a single file:
+        # it checks for all abums that have that file. The way it checks needs the
+        # [user-album] relationship because it'll return one less item. If the file
+        # has only 1 owner, it gets deleted for real. So the S3 item and metadata are
+        # gone, but it may have one more owner. When you do list files on THAT other
+        # owner, it won't be able to fetch metadata and your app will crash.
+
+        items = get_album_files(album_uuid)
+        for o in items:
+            album_item = dynamo_obj_to_python_obj(o)
+            item_uuid = album_item[TB_ALBUM_FILES_SK]
+            delete(username, album_uuid, item_uuid)
+
+        # Remove from user's list of albums.
     
-    # Remove each subfile: THIS HAS TO GO BEFORE BECAUSE:
-    # if you remove [user-album], then when you delete a single file:
-    # it checks for all abums that have that file. The way it checks needs the
-    # [user-album] relationship because it'll return one less item. If the file
-    # has only 1 owner, it gets deleted for real. So the S3 item and metadata are
-    # gone, but it may have one more owner. When you do list files on THAT other
-    # owner, it won't be able to fetch metadata and your app will crash.
-
-    items = get_album_files(album_uuid)
-    for o in items:
-        album_item = dynamo_obj_to_python_obj(o)
-        item_uuid = album_item[TB_ALBUM_FILES_SK]
-        delete(username, album_uuid, item_uuid)
-
-    # Remove from user's list of albums.
-  
-    dynamo_cli.delete_item(
-        TableName=TB_USER_ALBUMS_NAME,
-        Key=python_obj_to_dynamo_obj({
-            TB_USER_ALBUMS_PK: username,
-            TB_USER_ALBUMS_SK: album_uuid,
-        })
-    )
+        dynamo_cli.delete_item(
+            TableName=TB_USER_ALBUMS_NAME,
+            Key=python_obj_to_dynamo_obj({
+                TB_USER_ALBUMS_PK: username,
+                TB_USER_ALBUMS_SK: album_uuid,
+            })
+        )
     dynamo_cli.delete_item(
         TableName=TB_ALBUM_FILES_NAME,
         Key=python_obj_to_dynamo_obj({
